refactor: update ShoulderRocketLauncherClientSystem and SandevistanSystem for improved component handling and cleanup; remove unused ClearTargetsEvent
