--- Content.Shared/Imperial/XxRaay/Events/ClearTargetsEvent.cs
@@ -1,9 +1,0 @@
-using Robust.Shared.Serialization;
-
-namespace Content.Shared.Imperial.XxRaay.Events;
-
-[Serializable, NetSerializable]
-public sealed class ClearTargetsEvent : EntityEventArgs
-{
-}
-
